Move check new libraries task into a separate package
