+++ tests/tck-build-logic/src/main/groovy/org.graalvm.internal.tck-harness.gradle
@@ -16,8 +16,8 @@ import org.graalvm.internal.tck.ScaffoldTask
 import org.graalvm.internal.tck.GrypeTask
 import org.graalvm.internal.tck.TestedVersionUpdaterTask
 import org.graalvm.internal.tck.harness.tasks.CheckstyleInvocationTask
-import org.graalvm.internal.tck.harness.tasks.FetchExistingLibrariesWithNewerVersionsTask
-import org.graalvm.internal.tck.harness.tasks.GroupUnsupportedLibraries
+import org.graalvm.internal.tck.updaters.FetchExistingLibrariesWithNewerVersionsTask
+import org.graalvm.internal.tck.updaters.GroupUnsupportedLibraries
 import org.graalvm.internal.tck.harness.tasks.TestInvocationTask
