+++ tests/tck-build-logic/src/main/java/org/graalvm/internal/tck/ContributionTask.java
@@ -1,0 +1,2 @@
+package org.graalvm.internal.tck;public class ContributionTask {
+}
